+++ src/bitcoin-cli.cpp
@@ -6,8 +6,8 @@
 
 #include "chainparamsbase.h"
 #include "clientversion.h"
-#include "rpcclient.h"
-#include "rpcprotocol.h"
+#include "rpc/client.h"
+#include "rpc/protocol.h"
 #include "util.h"
 #include "sync.h"
 #include "utilstrencodings.h"
