+++ src/bitcoind.cpp
@@ -5,14 +5,13 @@
 
 #include "chainparams.h"
 #include "clientversion.h"
-#include "rpcserver.h"
+#include "rpc/server.h"
 #include "init.h"
 #include "noui.h"
 #include "scheduler.h"
 #include "util.h"
 #include "httpserver.h"
 #include "httprpc.h"
-#include "rpcserver.h"
 #include "unlimited.h"
 
 #include <boost/algorithm/string/predicate.hpp>
